feat: parse xml from json provided example
