+++ test-harness/specs/xml-response-AC-5.oas2.txt
@@ -1,0 +1,25 @@
+====test====
+Responding with an XML - invalid example.
+====spec====
+swagger: '2.0'
+paths:
+  "/pets":
+    get:
+      produces:
+        - application/xml
+      responses:
+        '200':
+          description: successful operation
+          examples:
+           application/xml:
+             id: 1
+             name: Sharik
+             status: available
+====server====
+mock -p 4010
+====command====
+curl -i http://127.0.0.1:4010/pets -H "Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8"
+====expect====
+HTTP/1.1 200 OK
+
+<id>1</id><name>Sharik</name><status>available</status>
