+++ test-harness/specs/xml-response-AC-6.oas2.txt
@@ -1,0 +1,42 @@
+====test====
+Responding with an XML.
+====spec====
+swagger: '2.0'
+paths:
+  "/pets":
+    get:
+      produces:
+        - application/json
+        - application/xml
+      responses:
+        '200':
+          description: successful operation
+          schema:
+            "$ref": "#/definitions/Pet"
+definitions:
+  Pet:
+    type: object
+    required:
+      - name
+    properties:
+      id:
+        type: integer
+        format: int64
+      name:
+        type: string
+        example: doggie
+      status:
+        type: string
+        description: pet status in the store
+        enum:
+          - available
+          - pending
+          - sold
+====server====
+mock -p 4010
+====command====
+curl -i http://127.0.0.1:4010/pets -H "Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8"
+====expect====
+HTTP/1.1 200 OK
+
+<id>0</id><name>doggie</name><status>available</status>
